Commit general, no se que le pasa la git, si esto anda me recibo

+++ workspace/tpa/src/main/java/com/tpa/app/Criterio.java
@@ -2,5 +2,5 @@ package com.tpa.app;
 
 public interface Criterio {
 	String nombre = "nombre";
-	int dameTuValor(Inscripcion inscripcion);
+	double dameTuValor(Inscripcion inscripcion);
 }
